패키지 분리 및 파일 이동, 리소스 추가, 바인딩 추가
strings.xml - 로그인 화면 리소스 추가
view 패키지 안에 activity, fragment 패키지로 분리
각 패키지로 파일 이동
MainViewActivity와 레이아웃 MainActivity, activity_main으로 변경
MainActivity와 레이아웃 LoginActivity, activity_login으로 변경
LoginActivity 뷰바인딩으로 변경, 회원가입으로 이동 추가
SplachActivity 소멸 애니메이션 제거
