+++ app/src/main/java/com/example/corner_library/view/fragment/PasswordInputFragment.kt
@@ -1,4 +1,4 @@
-package com.example.corner_library.view
+package com.example.corner_library.view.fragment
 
 import android.content.Context
 import android.os.Bundle
