+++ app/src/main/res/layout/activity_search_result.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.SearchResultActivity">
+    tools:context=".view.activity.SearchResultActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/search_result_toolbar"
