+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personal_color"
-    tools:context=".view.SplashActivity">
+    tools:context=".view.activity.SplashActivity">
 
     <ImageView
         android:id="@+id/splash_logo"
