+++ app/src/main/res/layout/fragment_password_input.xml
@@ -2,13 +2,13 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.NameInputFragment">
+    tools:context=".view.fragment.NameInputFragment">
 
     <data>
 
         <variable
             name="view"
-            type="com.example.corner_library.view.PasswordInputFragment" />
+            type="com.example.corner_library.view.fragment.PasswordInputFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
